UI FAB edit

+++ app/src/main/java/com/yoloswag/alex/seetitel/Whistle.java
@@ -17,5 +17,5 @@ public class Whistle {
     protected String audioURL;
     protected String fullText;
     protected int id;
-    protected int timestamp;
+    protected long timestamp;
 }
